1. На базе задания из занятия SRP расширить калькулятор [#219463]
