+++ chapter_008/src/main/java/ru/job4j/design/srp/report/ReportType.java
@@ -1,0 +1,5 @@
+package ru.job4j.design.srp.report;
+
+public enum ReportType {
+    STRING, HTML, XML, JSON, HR, ACCOUNTING
+}
